feat: upgrade set_decay_rate to u64 for finer-grained decay configuration
Upgrade set_decay_rate(numerator, denominator) from u32 to u64 params so
operators can express fine-grained λ ratios without hitting u32 overflow.
Updates constants, storage, events, parameter_governance, decay_fixed, and
get_decay_rate return type consistently. Adds NotInitialized and non-admin
rejection tests alongside existing validation coverage in test_decay.

+++ contracts/ledgerlens-score/src/constants.rs
@@ -65,10 +65,10 @@ pub const MAX_ADMIN_SIGNERS: u32 = 5;
 pub const DEFAULT_STALENESS_WINDOW_SECS: u64 = 604_800;
 pub const MAX_PAUSED_PAIRS: u32 = 50;
 pub const DECAY_FIXED_POINT_SCALE: u64 = 1_000_000;
-pub const DEFAULT_DECAY_LAMBDA_NUM: u32 = 0;
-pub const DEFAULT_DECAY_LAMBDA_DEN: u32 = 1;
-pub const MAX_DECAY_LAMBDA_NUM: u32 = 1;
-pub const MAX_DECAY_LAMBDA_DEN: u32 = 1;
+pub const DEFAULT_DECAY_LAMBDA_NUM: u64 = 0;
+pub const DEFAULT_DECAY_LAMBDA_DEN: u64 = 1;
+pub const MAX_DECAY_LAMBDA_NUM: u64 = 1;
+pub const MAX_DECAY_LAMBDA_DEN: u64 = 1;
 
 /// Minimum configurable escalation threshold (consecutive breaches).
 pub const MIN_ESCALATION_THRESHOLD: u32 = 1;
